Modificando_Codigo

+++ CONTROLADOR_TEMPERATURA_ARDUINO_DEF/lib/Proyectos_C/BCD_Conversion.cpp
@@ -7,7 +7,7 @@ BCD_Conversion::BCD_Conversion(){};
 int BCD_Conversion::Lectura_Centena(float Value) {
   /* code */
 //  float Temp=0;
-  Centena=(uint8_t)(Value/100);
+  Centena=(int)(Value/100);
   return Centena;
   //Decena=(uint8_t)((Value-Centena*100)/10);
   //Unidad=(uint8_t)((Value-Centena*100-Decena*10))
@@ -16,7 +16,7 @@ int BCD_Conversion::Lectura_Decena(float Value) {
   /* code */
 //  float Temp=0;
   //Centena=(uint8_t)(Value/100);
-  Decena=(uint8_t)((Value-Centena*100)/10);
+  Decena=(int)((Value-Centena*100)/10);
   return Decena;
   //Unidad=(uint8_t)((Value-Centena*100-Decena*10))
 }
@@ -26,6 +26,6 @@ int BCD_Conversion::Lectura_Unidad(float Value) {
 //  float Temp=0;
   //Centena=(uint8_t)(Value/100);
   //Decena=(uint8_t)((Value-Centena*100)/10);
-  Unidad=(uint8_t)((Value-Centena*100-Decena*10));
+  Unidad=(int)((Value-Centena*100-Decena*10));
   return Unidad;
 }
